+++ src/alignment/types/std_traits.rs
@@ -12,13 +12,6 @@ impl AsRef<[Ciglet]> for AlignmentStates {
     }
 }
 
-impl Default for AlignmentStates {
-    #[inline]
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
 impl PartialEq<Cigar> for AlignmentStates {
     #[inline]
     fn eq(&self, other: &Cigar) -> bool {
@@ -52,8 +45,12 @@ impl FromIterator<Ciglet> for Result<AlignmentStates, CigarError> {
     #[inline]
     fn from_iter<T: IntoIterator<Item = Ciglet>>(iter: T) -> Self {
         let mut last_op = 0;
+        let mut repeated_op = false;
 
-        let out = iter
+        // We do not expect repeated operations, so use collect and its
+        // specialized implementation relying on size_hint and/or in-place
+        // iteration
+        let mut out = iter
             .into_iter()
             .map(|ciglet| {
                 if ciglet.inc == 0 {
@@ -61,13 +58,34 @@ impl FromIterator<Ciglet> for Result<AlignmentStates, CigarError> {
                 } else if !is_valid_op(ciglet.op) {
                     Err(CigarError::InvalidOperation)
                 } else if ciglet.op == last_op {
-                    Err(CigarError::RepeatedOp)
+                    repeated_op = true;
+                    Ok(ciglet)
                 } else {
                     last_op = ciglet.op;
                     Ok(ciglet)
                 }
             })
             .collect::<Result<Vec<_>, _>>()?;
+
+        // Check whether a second pass for merging repeated operations is needed
+        if repeated_op {
+            // The original index in the output being read from
+            let mut original_index = 1;
+            // The current index being accumulated into
+            let mut merged_index = 0;
+            while let Some(ciglet) = out.get(original_index).copied() {
+                if ciglet.op == out[merged_index].op {
+                    out[merged_index].inc += ciglet.inc;
+                } else {
+                    merged_index += 1;
+                    out[merged_index] = out[original_index];
+                }
+                original_index += 1;
+            }
+            // We have written to merged_index, so we must keep merged_index+1
+            // length
+            out.truncate(merged_index + 1);
+        }
 
         Ok(AlignmentStates(out))
     }
@@ -93,15 +111,37 @@ impl std::fmt::Display for AlignmentStates {
 impl TryFrom<&[u8]> for AlignmentStates {
     type Error = CigarError;
 
-    fn try_from(v: &[u8]) -> Result<Self, Self::Error> {
+    fn try_from(v: &[u8]) -> Result<Self, CigarError> {
+        let mut states = AlignmentStates::new();
+
         if v == b"*" {
-            Ok(AlignmentStates::new())
+            Ok(states)
         } else {
-            // Validity: CigletIteratorChecked checks for increments that are
-            // zero and adjacent operations which are the same
-            CigletIteratorChecked::new(v)
-                .collect::<Result<Vec<_>, _>>()
-                .map(AlignmentStates)
+            let mut iter = CigletIteratorChecked::new(v);
+            let Some(prev_ciglet) = iter.next() else {
+                return Ok(states);
+            };
+            let mut prev_ciglet = prev_ciglet?;
+
+            // More efficient version of add_ciglet: we don't need to check
+            // whether a previous ciglet exists, nor do we have to check for
+            // non-zero increments
+            for ciglet in iter {
+                let ciglet = ciglet?;
+                if prev_ciglet.op == ciglet.op {
+                    prev_ciglet.inc += ciglet.inc;
+                } else {
+                    // Validity: CigletIteratorChecked errors on zero
+                    // increments, and we have ensured the operation is
+                    // different from the previous
+                    states.0.push(prev_ciglet);
+                    prev_ciglet = ciglet;
+                }
+            }
+
+            // Push the last ciglet
+            states.0.push(prev_ciglet);
+            Ok(states)
         }
     }
 }